Function-per-file style
