+++ src/utils/formatClassName.js
@@ -1,2 +1,3 @@
 const formatClassName = (array = []) => array.filter(Boolean).join(" ");
+
 export default formatClassName;
