+++ src/utils/hexToRgb.js
@@ -1,0 +1,19 @@
+const hexToRgb = (hex) => {
+  if (hex[0] === "#") hex = hex.substr(1);
+
+  if (hex.length === 3) {
+    return {
+      r: parseInt(hex[0] + hex[0], 16),
+      g: parseInt(hex[1] + hex[1], 16),
+      b: parseInt(hex[2] + hex[2], 16),
+    };
+  }
+
+  return {
+    r: parseInt(hex.substr(0, 2), 16),
+    g: parseInt(hex.substr(2, 2), 16),
+    b: parseInt(hex.substr(4, 2), 16),
+  };
+};
+
+export default hexToRgb;
